fix unit tests

+++ api/observer_test.go
@@ -22,15 +22,15 @@ func Test_parseSubscriptions(t *testing.T) {
 			webhook: "http://127.0.0.1:8080",
 			wantSubs: []interface{}{
 				&storage.Subscription{
-					Coin: uint(2), Webhook: "http://127.0.0.1:8080",
+					Coin: 2, Webhook: "http://127.0.0.1:8080",
 					Address: "zpub6rH4MwgyTmuexAX6HAraks5cKv5BbtmwdLirvnU5845ovUJb4abgjt9DtXK4ZEaToRrNj8dQznuLC6Nka4eMviGMinCVMUxKLpuyddcG9Vc",
 				},
 				&storage.Subscription{
-					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
+					Coin: 0, Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6BpYi6J1GZzfY3yY7DbhLLccF3efQa18nQngM3jaehgtNSoEgk6UtPULpC3oK5oA3trczY8Ld34LFw1USMPfGHwTEizdD5QyGcMyuh2UoBA",
 				},
 				&storage.Subscription{
-					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
+					Coin: 0, Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6CYwPfnPJLPquufPkb98coSb3mdy1CgaZrWUtYWGJTJ4VWZUbzH9HLGy7nHpP7DG4UdTkYYpirkTWQSP7pWHsrk24Nos5oYNHpfr4BgPVTL",
 				},
 			},
@@ -42,11 +42,11 @@ func Test_parseSubscriptions(t *testing.T) {
 			webhook: "http://127.0.0.1:8080",
 			wantSubs: []interface{}{
 				&storage.Subscription{
-					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
+					Coin: 0, Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6BpYi6J1GZzfY3yY7DbhLLccF3efQa18nQngM3jaehgtNSoEgk6UtPULpC3oK5oA3trczY8Ld34LFw1USMPfGHwTEizdD5QyGcMyuh2UoBA",
 				},
 				&storage.Subscription{
-					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
+					Coin: 0, Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6CYwPfnPJLPquufPkb98coSb3mdy1CgaZrWUtYWGJTJ4VWZUbzH9HLGy7nHpP7DG4UdTkYYpirkTWQSP7pWHsrk24Nos5oYNHpfr4BgPVTL",
 				},
 			},
